# specs/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Tax Dividend AI Backend Internal API
  version: 1.0.0
  description: Internal API for Backend Services
servers:
  - url: http://localhost:8081/internal
    description: Local development server
components:
  schemas:
...
    DividendStatement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        sourceFileName:
          type: string
          description: Original filename of uploaded broker statement
        sourceFileS3Key:
          type: string
          description: S3/MinIO key for stored file
        broker:
          type: string
          description: Broker name (InteractiveBrokers, Swissquote, etc.)
        periodStart:
          type: string
          format: date
          description: Start of period covered by statement
        periodEnd:
          type: string
          format: date
          description: End of period covered by statement
        status:
          $ref: '#/components/schemas/DividendStatementStatus'
        parsedAt:
          type: string
          format: date-time
          nullable: true
        validatedAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        sentMethod:
          type: string
          nullable: true
          description: Submission method (EMAIL, POSTAL, ONLINE)
        sentNotes:
          type: string
          nullable: true
          description: User notes about submission
        paidAt:
          type: string
          format: date-time
          nullable: true
        paidAmount:
          type: number
          nullable: true
          description: Amount received from tax authority
        parsedBy:
          type: string
          nullable: true
          description: Who/what triggered parsing (AI_AGENT, MANUAL, etc.)
        dividendCount:
          type: integer
          description: Number of dividends extracted from statement
        totalGrossAmount:
          type: number
          description: Total gross amount from all dividends
        totalReclaimable:
          type: number
          description: Total reclaimable amount from all dividends
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DividendStatementStatus:
      type: string
      enum:
        - UPLOADED
        - PARSING
        - PARSED
        - VALIDATED
        - SENT
        - PAID
      description: |
        Status workflow for dividend statements:
        - UPLOADED: File uploaded, awaiting parsing
        - PARSING: Currently being parsed by AI Agent
        - PARSED: Parsing complete, dividends extracted
        - VALIDATED: Dividends validated, forms generated and downloaded
        - SENT: User has submitted forms to tax authority (offline)
        - PAID: Reimbursement received (marked manually by user)
    DividendStatementUpdateDTO:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/DividendStatementStatus'
        sentMethod:
          type: string
          description: For SENT status - submission method
        sentNotes:
          type: string
          description: For SENT status - additional notes
        paidAmount:
          type: number
          description: For PAID status - amount received
        paidAt:
          type: string
          format: date-time
          description: For PAID status - payment date (defaults to now if not provided)
        parsedBy:
          type: string
          description: For AI Agent updates - who triggered the parsing

    # Form schemas
...
paths:
...
  /dividend-statements:
    post:
      tags:
        - DividendStatements
      summary: Upload broker statement
      description: Upload a broker statement PDF/CSV file. Creates a new DividendStatement with status UPLOADED.
      operationId: uploadDividendStatement
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: broker
          in: query
          required: false
          description: Broker name (InteractiveBrokers, Swissquote, etc.)
          schema:
            type: string
        - name: periodStart
          in: query
          required: false
          description: Start of period covered by statement
          schema:
            type: string
            format: date
        - name: periodEnd
          in: query
          required: false
          description: End of period covered by statement
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Broker statement file (PDF or CSV)
      responses:
        '201':
          description: Statement uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DividendStatement'
        '400':
          description: Bad request - invalid file or parameters

    get:
      tags:
        - DividendStatements
      summary: List user's dividend statements
      description: List dividend statements with optional status filter and pagination
      operationId: listDividendStatements
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          description: Filter by status
          schema:
            $ref: '#/components/schemas/DividendStatementStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Page of dividend statements
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/DividendStatement'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
  /dividend-statements/{id}:
    get:
      tags:
        - DividendStatements
      summary: Get dividend statement by ID
      description: Get a specific dividend statement. Validates ownership.
      operationId: getDividendStatement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dividend statement details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DividendStatement'
        '404':
          description: Not found or not owned by user

    patch:
      tags:
        - DividendStatements
      summary: Update dividend statement status
      description: |
        Update statement status. Validates status transitions:
        - UPLOADED → PARSING (AI Agent starts parsing)
        - PARSING → PARSED (AI Agent completes parsing)
        - PARSED → VALIDATED (User downloads forms)
        - VALIDATED → SENT (User submits forms offline)
        - SENT → PAID (User receives payment)
      operationId: updateDividendStatementStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DividendStatementUpdateDTO'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DividendStatement'
        '400':
          description: Invalid status transition
        '404':
          description: Not found or not owned by user

    delete:
      tags:
        - DividendStatements
      summary: Delete dividend statement
      description: Delete statement and associated file from storage
      operationId: deleteDividendStatement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted successfully
        '404':
          description: Not found or not owned by user
  /dividend-statements/by-date-range:
    get:
      tags:
        - DividendStatements
      summary: Find statements by date range
      description: Find statements whose period overlaps with the specified date range
      operationId: getDividendStatementsByDateRange
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of statements in date range
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DividendStatement'
  /dividend-statements/count-by-status:
    get:
      tags:
        - DividendStatements
      summary: Count statements by status
      description: Get count of statements for a specific status
      operationId: countDividendStatementsByStatus
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/DividendStatementStatus'
      responses:
        '200':
          description: Count of statements
          content:
            application/json:
              schema:
                type: integer
                format: int64

  # Forms endpoints
...